refactor(resolution): move LocalFSC into localres.py alongside LocalResolution
LocalFSC and LocalResolution both produce local (spatially-resolved)
resolution *maps* — LocalFSC via two-half-map block FSC/FRC (blocres-style),
LocalResolution via a single-map ResMap-style statistical test — yet one lived
in FSC.py (among the global one-number FSC/SSNR/plot classes) and the other in
localres.py. That split obscured which is which.

Co-locate them: FSC.py is now purely global resolution (FourierShellCorr,
FSCPlot/FRCPlot/SSNRPlot, RandomFSC); localres.py is the home for the local
resolution maps. Pure move, no behaviour change.

- LocalFSC moved verbatim from FSC.py to localres.py; it is self-contained
  (references none of the global FSC classes), so no cross-import is needed.
- localres.py gains the imports LocalFSC uses (itertools, scipy.fft fftn/fftfreq
  aliases, RegularGridInterpolator); show_resolution_map was already imported.
- __all__ updated on both files; module docstring of localres.py rewritten to
  describe both estimators and when to use each.
- Public API unchanged: toupy.resolution.LocalFSC still resolves (flat import *).
  Verified by importing the package and running LocalFSC end-to-end on a 24^3
  pair of half-maps.

Class names kept as-is to avoid an API break.
