+++ toupy/resolution/localres.py
@@ -1,12 +1,23 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 """
-Single-map local resolution estimation via a ResMap-inspired statistical test.
+Local (spatially-resolved) resolution estimation.
+
+Two complementary estimators live here:
+
+* ``LocalResolution`` -- single-map, ResMap-style statistical test; needs
+  only one reconstruction (usable for local tomography).
+* ``LocalFSC`` -- two-half-map, block-wise FSC/FRC (blocres-style); needs a
+  pair of independent half-reconstructions.
+
+Global (single-number) FSC/SSNR live in ``FSC.py``; this module is the home
+for the local resolution *maps*.
 """
 
 # standard library
 import time
 import warnings
+import itertools
 
 # third party
 import numpy as np
@@ -18,12 +29,14 @@ from scipy.ndimage import (
     generate_binary_structure,
 )
 from scipy.stats import norm as _norm
+from scipy.fft import fftn as _fftn, fftfreq as _fftfreq
+from scipy.interpolate import RegularGridInterpolator
 
 # local
 from ..utils import tqdm
 from ..utils.plot_utils import show_resolution_map
 
-__all__ = ["LocalResolution"]
+__all__ = ["LocalResolution", "LocalFSC"]
 
 
 class LocalResolution:
@@ -690,6 +703,384 @@ class LocalResolution:
             self.vol.ndim,
             title="LocalResolution map",
             filename="LocalResolution_map.png",
+            slice_idx=slice_idx,
+            axis=axis,
+            cmap=cmap,
+            vmin=vmin,
+            vmax=vmax,
+        )
+        return self.resolution_map
+
+
+# ---------------------------------------------------------------------------
+# LocalFSC  (block-wise FSC/FRC; moved here from FSC.py so both local
+# resolution-map estimators live together)
+# ---------------------------------------------------------------------------
+
+class LocalFSC:
+    """
+    Local resolution estimation via block-wise FSC (3-D) or FRC (2-D).
+
+    Divides the volume into overlapping boxes, computes an independent
+    Fourier Shell/Ring Correlation within each box, and assembles a
+    spatial map of local resolution interpolated to the input shape.
+
+    Parameters
+    ----------
+    vol1 : ndarray
+        First independent half-reconstruction (2-D or 3-D).
+    vol2 : ndarray
+        Second independent half-reconstruction, same shape as *vol1*.
+    pixel_size : float, optional
+        Physical voxel size (any consistent unit). Default ``1.0``.
+    box_size : int, optional
+        Side length (in voxels) of the cubic (or square) analysis box.
+        Smaller boxes give finer spatial sampling but noisier estimates.
+        Default ``32``.
+    step : int or None, optional
+        Step size in voxels between adjacent box centres.
+        Default ``box_size // 2`` (50 % overlap).
+    threshold : float or {'1bit', 'halfbit'}, optional
+        Criterion used to define the local resolution limit.
+
+        float
+            A fixed FSC threshold applied at every shell, e.g. ``0.143``
+            (the Rosenthal & Henderson, 2003, "gold-standard" criterion —
+            *not* the half-bit criterion, despite the common conflation
+            of the two [3]_) or ``0.5`` (the fixed threshold used by
+            Cardone, Heymann & Steven, 2013, in ``blocres`` [1]_, the
+            reference local-FSC implementation this class follows).
+            Default ``0.143``.
+        ``'1bit'`` or ``'halfbit'``
+            The shell-size-corrected one-bit or half-bit threshold of
+            van Heel & Schatz, 2005 [2]_, evaluated per shell from the
+            number of independent Fourier voxels actually present in
+            *this* box, rather than the fixed asymptotic value. Rohou,
+            2020 [4]_, shows that fixed FSC thresholds (0.143 or 0.5
+            alike) are biased toward over-estimating resolution in the
+            sparsely populated low-frequency shells of small
+            sub-volumes — precisely the regime ``LocalFSC`` operates in
+            — so this option is recommended whenever ``box_size`` is
+            small relative to the reconstruction.
+
+    Attributes
+    ----------
+    resolution_map : ndarray
+        Local full-period resolution in pixels (van Heel convention),
+        same shape as *vol1*.
+    resolution_map_phys : ndarray
+        Local full-period resolution in physical units
+        (``resolution_map * pixel_size``).
+    resolution_map_half : ndarray
+        Local half-period resolution in pixels (Rayleigh convention):
+        ``resolution_map / 2``.
+    resolution_map_phys_half : ndarray
+        Local half-period resolution in physical units:
+        ``resolution_map_phys / 2``.
+    resolution_median : float
+        Median full-period local resolution in pixels.
+    resolution_mean : float
+        Mean full-period local resolution in pixels.
+    resolution_std : float
+        Standard deviation of full-period local resolution in pixels.
+    resolution_median_half : float
+        Median half-period local resolution in pixels.
+    resolution_mean_half : float
+        Mean half-period local resolution in pixels.
+    resolution_std_half : float
+        Standard deviation of half-period local resolution in pixels.
+
+    References
+    ----------
+    .. [1] G. Cardone, J. B. Heymann, and A. C. Steven, "One number does
+       not fit all: Mapping local variations in resolution in cryo-EM
+       reconstructions", J. Struct. Biol. 184, 226-236 (2013).
+       https://doi.org/10.1016/j.jsb.2013.08.002
+    .. [2] M. van Heel and M. Schatz, "Fourier shell correlation threshold
+       criteria", J. Struct. Biol. 151, 250-262 (2005).
+       https://doi.org/10.1016/j.jsb.2005.05.009
+    .. [3] P. B. Rosenthal and R. Henderson, "Optimal determination of
+       particle orientation, absolute hand, and contrast loss in
+       single-particle electron cryomicroscopy", J. Mol. Biol. 333,
+       721-745 (2003). https://doi.org/10.1016/j.jmb.2003.07.013
+    .. [4] A. Rohou, "Fourier shell correlation criteria for local
+       resolution estimation", bioRxiv 2020.03.01.972067 (2020).
+       https://doi.org/10.1101/2020.03.01.972067
+    """
+
+    _ADAPTIVE_SNR = {
+        "1bit": 0.5,
+        "one-bit": 0.5,
+        "onebit": 0.5,
+        "halfbit": 0.2071,
+        "half-bit": 0.2071,
+        "1/2bit": 0.2071,
+    }
+
+    def __init__(self, vol1, vol2, pixel_size=1.0, box_size=32, step=None, threshold=0.143):
+        print("Calling the class LocalFSC")
+        vol1 = np.asarray(vol1, dtype=np.float64)
+        vol2 = np.asarray(vol2, dtype=np.float64)
+        if vol1.shape != vol2.shape:
+            raise ValueError("LocalFSC: vol1 and vol2 must have the same shape.")
+        if vol1.ndim not in (2, 3):
+            raise ValueError("LocalFSC: input must be 2-D or 3-D.")
+        self.vol1 = vol1
+        self.vol2 = vol2
+        self.shape = vol1.shape
+        self.ndim = vol1.ndim
+        self.pixel_size = float(pixel_size)
+        self.box_size = int(box_size)
+        self.step = int(step) if step is not None else self.box_size // 2
+
+        # threshold: either a fixed FSC value (as before), or a
+        # shell-size-corrected criterion ('1bit' / 'halfbit', van Heel &
+        # Schatz, 2005) evaluated per shell from the number of
+        # independent Fourier voxels n(xi) actually present in that box.
+        # See the class docstring for the rationale (Rohou, 2020).
+        if isinstance(threshold, str):
+            key = threshold.lower()
+            if key not in self._ADAPTIVE_SNR:
+                raise ValueError(
+                    f"LocalFSC: threshold string must be one of "
+                    f"{sorted(set(self._ADAPTIVE_SNR))!r}, got {threshold!r}."
+                )
+            self.threshold = threshold
+            self._adaptive_snr = self._ADAPTIVE_SNR[key]
+        else:
+            self.threshold = float(threshold)
+            self._adaptive_snr = None
+
+        self._compute()
+
+        ps = self.pixel_size
+        print(f"  LocalFSC median resolution (full period) : "
+              f"{self.resolution_median:.2f} px "
+              f"({self.resolution_median * ps:.4g} physical units)")
+        print(f"  LocalFSC median resolution (half period) : "
+              f"{self.resolution_median_half:.2f} px "
+              f"({self.resolution_median_half * ps:.4g} physical units)")
+        print(f"  LocalFSC mean   resolution (full period) : "
+              f"{self.resolution_mean:.2f} ± {self.resolution_std:.2f} px")
+
+    def _make_window(self):
+        """
+        Build a Hanning window matching the analysis box shape.
+
+        Returns
+        -------
+        window : ndarray
+            Hanning window of shape ``(box_size,) * ndim``.
+        """
+        w1d = np.hanning(self.box_size)
+        window = w1d.copy()
+        for _ in range(self.ndim - 1):
+            window = np.multiply.outer(window, w1d)
+        return window
+
+    def _fsc_box(self, box1, box2, window):
+        """
+        Compute local resolution (pixels) from a pair of overlapping boxes.
+
+        Subtracts the mean, applies *window*, computes the FFT, builds
+        radial shells using ``scipy.fft.fftfreq``, and returns the local
+        resolution at the shell where FSC first drops below the chosen
+        threshold — either the fixed value ``self.threshold``, or, when
+        ``self._adaptive_snr`` is set, the van Heel & Schatz (2005)
+        shell-size-corrected threshold evaluated from the actual number
+        of independent Fourier voxels in each shell of *this* box.
+
+        Parameters
+        ----------
+        box1 : ndarray
+            First analysis box (shape ``(box_size,) * ndim``).
+        box2 : ndarray
+            Second analysis box, same shape as *box1*.
+        window : ndarray
+            Hanning window of shape ``(box_size,) * ndim``.
+
+        Returns
+        -------
+        resolution_px : float
+            Local resolution in pixels.  Returns ``box_size * 2.0`` when
+            the FSC never reaches the chosen threshold (bad estimate).
+        """
+        b = self.box_size
+        eps = np.finfo(np.float64).tiny
+
+        b1 = (box1 - box1.mean()) * window
+        b2 = (box2 - box2.mean()) * window
+
+        F1 = _fftn(b1)
+        F2 = _fftn(b2)
+
+        # Build radial coordinate grid in cycles/pixel (FFT layout)
+        freqs = [_fftfreq(b) for _ in range(self.ndim)]
+        grids = np.meshgrid(*freqs, indexing='ij')
+        R = np.sqrt(sum(g ** 2 for g in grids))
+
+        # Radial shells: integer bins in units of 1/box_size
+        shell_idx = np.round(R * b).astype(np.int32)
+        max_shell = b // 2
+
+        fsc_vals = []
+        n_vals = []
+        for s in range(max_shell + 1):
+            mask = shell_idx == s
+            n_shell = int(np.sum(mask))
+            n_vals.append(n_shell)
+            if n_shell == 0:
+                fsc_vals.append(0.0)
+                continue
+            f1 = F1[mask]
+            f2 = F2[mask]
+            cross = np.abs(np.vdot(f2, f1))
+            a1 = np.abs(np.vdot(f1, f1))
+            a2 = np.abs(np.vdot(f2, f2))
+            denom = np.sqrt(a1 * a2)
+            fsc_vals.append(float(cross / denom) if denom > eps else 0.0)
+
+        fsc_arr = np.array(fsc_vals)
+
+        if self._adaptive_snr is not None:
+            # Shell-size-corrected threshold (van Heel & Schatz, 2005),
+            # evaluated per shell from n(xi) actually present in this box:
+            #   T(xi) = [SNR + (2*sqrt(SNR)+1)/sqrt(n)] / [SNR + 1 + 2*sqrt(SNR)/sqrt(n)]
+            n_arr = np.maximum(np.array(n_vals, dtype=np.float64), 1.0)
+            snr = self._adaptive_snr
+            sqrt_snr = np.sqrt(snr)
+            sqrt_n = np.sqrt(n_arr)
+            thr_arr = (snr + (2.0 * sqrt_snr + 1.0) / sqrt_n) / (
+                snr + 1.0 + 2.0 * sqrt_snr / sqrt_n
+            )
+            above = np.where(fsc_arr[1:] >= thr_arr[1:])[0] + 1  # skip DC
+        else:
+            # Find the highest shell where FSC >= fixed threshold
+            above = np.where(fsc_arr[1:] >= self.threshold)[0] + 1  # skip DC
+
+        if len(above) == 0:
+            return float(b * 2.0)
+        r_res_shell = above[-1]
+        # Convert shell index to resolution in pixels: shell s = s/b cycles/pixel
+        r_cpx = r_res_shell / float(b)
+        if r_cpx < eps:
+            return float(b * 2.0)
+        return 1.0 / r_cpx
+
+    def _compute(self):
+        """
+        Iterate over the coarse grid, compute local FSC per box, and
+        interpolate the result to the full volume shape.
+
+        Stores
+        ------
+        resolution_map : ndarray
+        resolution_map_phys : ndarray
+        resolution_median : float
+        resolution_mean : float
+        resolution_std : float
+        """
+        vol1 = self.vol1
+        vol2 = self.vol2
+        shape = self.shape
+        ndim = self.ndim
+        b = self.box_size
+        step = self.step
+        half = b // 2
+        window = self._make_window()
+
+        # Coarse grid: centre positions
+        axes = [np.arange(half, shape[d] - half, step) for d in range(ndim)]
+
+        for d, ax in enumerate(axes):
+            if len(ax) < 2:
+                warnings.warn(
+                    f"LocalFSC: axis {d} has fewer than 2 grid points "
+                    f"(box_size={b} may be too large for dimension {shape[d]}).  "
+                    "Resolution map may be unreliable.",
+                    UserWarning,
+                    stacklevel=2,
+                )
+
+        coarse_shape = tuple(len(ax) for ax in axes)
+        coarse_map = np.empty(coarse_shape, dtype=np.float64)
+
+        idx_ranges = [range(len(ax)) for ax in axes]
+        for grid_idx in itertools.product(*idx_ranges):
+            centres = tuple(int(axes[d][grid_idx[d]]) for d in range(ndim))
+            slices = tuple(
+                slice(max(0, centres[d] - half), min(shape[d], centres[d] + half))
+                for d in range(ndim)
+            )
+            box1 = vol1[slices]
+            box2 = vol2[slices]
+
+            # Pad to box_size if the box is smaller near the border
+            if box1.shape != (b,) * ndim:
+                pad1 = [(0, b - box1.shape[dd]) for dd in range(ndim)]
+                box1 = np.pad(box1, pad1)
+                box2 = np.pad(box2, pad1)
+
+            coarse_map[grid_idx] = self._fsc_box(box1, box2, window)
+
+        # Interpolate to full volume using RegularGridInterpolator
+        interp = RegularGridInterpolator(
+            axes,
+            coarse_map,
+            method='linear',
+            bounds_error=False,
+            fill_value=None,
+        )
+
+        # Build full meshgrid of integer coordinates
+        full_grids = np.mgrid[tuple(slice(0, shape[d]) for d in range(ndim))]
+        # full_grids shape: (ndim, *shape)
+        pts = np.stack([full_grids[d].ravel() for d in range(ndim)], axis=-1)
+        res_full = interp(pts).reshape(shape)
+
+        self.resolution_map       = res_full
+        self.resolution_map_phys  = res_full * self.pixel_size
+        self.resolution_map_half  = res_full / 2.0
+        self.resolution_map_phys_half = self.resolution_map_phys / 2.0
+
+        self.resolution_median      = float(np.median(res_full))
+        self.resolution_mean        = float(np.mean(res_full))
+        self.resolution_std         = float(np.std(res_full))
+        self.resolution_median_half = self.resolution_median / 2.0
+        self.resolution_mean_half   = self.resolution_mean   / 2.0
+        self.resolution_std_half    = self.resolution_std    / 2.0
+
+    def plot(self, slice_idx=None, axis=0, cmap='viridis_r', vmin=None, vmax=None):
+        """
+        Display the local resolution map and save it to ``LocalFSC_resmap.png``.
+
+        For 3-D volumes the central slice (or the slice specified by
+        *slice_idx*) along *axis* is shown.  The colour bar is in pixels.
+
+        Parameters
+        ----------
+        slice_idx : int or None, optional
+            Slice index along *axis* to display.  Defaults to the central
+            slice.  Ignored for 2-D input.
+        axis : int, optional
+            Volume axis along which to slice.  Default ``0``.
+        cmap : str, optional
+            Matplotlib colourmap.  Default ``'viridis_r'``.
+        vmin : float or None, optional
+            Lower colour-scale limit.  ``None`` uses the data minimum.
+        vmax : float or None, optional
+            Upper colour-scale limit.  ``None`` uses the data maximum.
+
+        Returns
+        -------
+        resolution_map : ndarray
+            The full local-resolution map (same shape as the input volume).
+        """
+        show_resolution_map(
+            self.resolution_map,
+            self.ndim,
+            title="LocalFSC resolution map",
+            filename="LocalFSC_resmap.png",
             slice_idx=slice_idx,
             axis=axis,
             cmap=cmap,
